+++ src/main/java/racingcar/domain/cars/factory/CarsFactory.java
@@ -1,8 +1,9 @@
-package racingcar.domain.cars;
+package racingcar.domain.cars.factory;
 
 import java.util.ArrayList;
 import java.util.List;
 import racingcar.domain.car.Car;
+import racingcar.domain.cars.Cars;
 
 public class CarsFactory {
     public static Cars fromNames(String[] names) {
