+++ src/main/java/racingcar/domain/cars/filter/CarsFilter.java
@@ -1,7 +1,8 @@
-package racingcar.domain.cars;
+package racingcar.domain.cars.filter;
 
 import java.util.List;
 import racingcar.domain.car.Car;
+import racingcar.domain.cars.Cars;
 
 public class CarsFilter {
     public static Cars maxMoved(List<Car> carList) {
