+++ src/main/java/racingcar/game/view/InputView.java
@@ -2,7 +2,7 @@ package racingcar.game.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import racingcar.domain.cars.Cars;
-import racingcar.domain.cars.CarsFactory;
+import racingcar.domain.cars.factory.CarsFactory;
 import racingcar.domain.race.RaceCount;
 
 public class InputView {
